refactor: :fire: refactore file structure for simplicity
